lab1-mcp-finance-lab/pyproject.toml
[project]
name = "mcp-finance"
version = "0.1.0"
description = "MCP сервер для финансовых расчетов: кредиты, вклады и инвестиции"
authors = [
    {name = "ermadatov", email = "ermadatov@cloud.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastmcp>=2.10.0,<3.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "prometheus-client",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp-proto-http",
    "python-dotenv",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
